Add power automatically find the latest serial port and prompt.
Add help instruction information or board prompt information.

Add runfile to run the local code, which is a combination of put + execfile.

Add execfile to run the local code, which is a combination of put + repl.

Fix exec file feedback running process information or error code, and support Ctrl + C way to stop the program.

Add an open reconnect mechanism that automatically reconnects the board.

The reset mechanism under Darwin system was tested to be invalid, so special case shielding was performed.